Created README.md and deleted work-in-progress code snippets

+++ components/EventList.js
@@ -1,31 +1,12 @@
 import { format } from 'date-fns';
 
 export default function EventList({ events }) {
-  // const handleDelete = async (item) => {
-  //   const answer = window.confirm(
-  //     `Do you really want to delete the event`
-  //   //   : ${item.name} on ${format(
-  //   //     new Date(item.datetime),
-  //   //     'eee., dd.MM.yyyy, HH:mm,',
-  //   //   )} from your list?`,
-  //   // );
-
-  //   if (answer === true) {
-  //     console.log(item);
-  //     await fetch(`http://localhost:3000/events/${item.id}`, {
-  //       method: 'DELETE',
-  //     });
-
-  //     // This is just a fast way of refreshing the information
-  //     window.location.reload();
-  //   }
-  // };
   return (
     <div>
       <h3>All Events</h3>
       <ol>
         {events.map((event) => (
-          <li key={event.id}>
+          <li key={event.eventsId}>
             <div>
               <p>{`${format(
                 new Date(event.datetime),
@@ -33,24 +14,23 @@ export default function EventList({ events }) {
               )} ${event.sportId}, ${event.homeTeamId}-${event.awayTeamId}, ${
                 event.details
               } `}</p>
+
               <button
-                // onClick={() => handleDelete()}
                 onClick={async () => {
-                  const response = await fetch(
-                    `http://localhost:3000/api/events`,
-                    {
+                  const answer = window.confirm(
+                    `Do you really want to delete the event`,
+                  );
+                  if (answer === true) {
+                    await fetch(`http://localhost:3000/api/events`, {
                       method: 'DELETE',
                       headers: {
                         'Content-Type': 'application/json',
                       },
                       body: JSON.stringify({ id: event.eventsId }),
-                    },
-                  );
+                    });
 
-                  if (response.status === 200) {
-                    alert('Sikeresen törlődött');
+                    window.location.reload();
                   }
-                  window.location.reload();
                 }}
               >
                 Delete
